Currency rate should be double

+++ database/migrations/2023_12_22_112848_create_currencies_table.php
@@ -12,7 +12,7 @@ return new class extends Migration {
     {
         Schema::create('currencies', function (Blueprint $table) {
             $table->string('symbol')->primary();
-            $table->decimal('rate', 16, 4);
+            $table->double('rate');
             $table->timestamps();
         });
     }
